+++ src/pool/cas.rs
@@ -5,12 +5,10 @@
 
 use core::{
     cell::UnsafeCell,
-    convert::TryFrom,
     marker::PhantomData,
-    mem,
-    num::NonZeroUsize,
+    num::{NonZeroU32, NonZeroU64},
     ptr::NonNull,
-    sync::atomic::{AtomicUsize, Ordering},
+    sync::atomic::{AtomicU64, Ordering},
 };
 
 /// Unfortunate implementation detail required to use the
@@ -88,14 +86,33 @@ impl<T> Stack<T> {
     }
 }
 
-fn anchor<T>() -> *mut T {
-    static mut ANCHOR: u8 = 0;
-    (unsafe { &mut ANCHOR } as *mut u8 as usize & !(mem::align_of::<T>() - 1)) as *mut T
+#[cfg(target_arch = "x86_64")]
+fn anchor<T>(init: Option<*mut T>) -> *mut T {
+    use core::sync::atomic::AtomicU8;
+
+    use spin::Once;
+
+    static LAZY_ANCHOR: Once<usize> = Once::new();
+
+    let likely_unaligned_address = if let Some(init) = init {
+        *LAZY_ANCHOR.call_once(|| init as usize)
+    } else {
+        LAZY_ANCHOR.get().copied().unwrap_or_else(|| {
+            // we may hit this branch with Pool of ZSTs where `grow` does not need to be called
+            static BSS_ANCHOR: AtomicU8 = AtomicU8::new(0);
+            &BSS_ANCHOR as *const _ as usize
+        })
+    };
+
+    let alignment_mask = !(core::mem::align_of::<T>() - 1);
+    let well_aligned_address = likely_unaligned_address & alignment_mask;
+    well_aligned_address as *mut T
 }
 
-/// Anchored pointer. This is a (signed) 32-bit offset from `anchor` plus a 32-bit tag
+/// On x86_64, anchored pointer. This is a (signed) 32-bit offset from `anchor` plus a 32-bit tag
+/// On x86, this is a pointer plus a 32-bit tag
 pub struct Ptr<T> {
-    inner: NonZeroUsize,
+    inner: NonZeroU64,
     _marker: PhantomData<*mut T>,
 }
 
@@ -107,37 +124,52 @@ impl<T> Clone for Ptr<T> {
 
 impl<T> Copy for Ptr<T> {}
 
+fn initial_tag_value() -> NonZeroU32 {
+    NonZeroU32::new(1).unwrap()
+}
+
 impl<T> Ptr<T> {
+    #[cfg(target_arch = "x86_64")]
     pub fn new(p: *mut T) -> Option<Self> {
-        i32::try_from((p as isize).wrapping_sub(anchor::<T>() as isize))
+        use core::convert::TryFrom;
+
+        i32::try_from((p as isize).wrapping_sub(anchor::<T>(Some(p)) as isize))
             .ok()
-            .map(|offset| unsafe { Ptr::from_parts(0, offset) })
+            .map(|offset| unsafe { Ptr::from_parts(initial_tag_value(), offset) })
     }
 
-    unsafe fn from_parts(tag: u32, offset: i32) -> Self {
+    #[cfg(target_arch = "x86")]
+    pub fn new(p: *mut T) -> Option<Self> {
+        Some(unsafe { Ptr::from_parts(initial_tag_value(), p as i32) })
+    }
+
+    unsafe fn from_parts(tag: NonZeroU32, offset: i32) -> Self {
         Self {
-            inner: NonZeroUsize::new_unchecked((tag as usize) << 32 | (offset as u32 as usize)),
+            inner: NonZeroU64::new_unchecked((tag.get() as u64) << 32 | (offset as u32 as u64)),
             _marker: PhantomData,
         }
     }
 
-    fn from_usize(p: usize) -> Option<Self> {
-        NonZeroUsize::new(p).map(|inner| Self {
+    fn from_u64(p: u64) -> Option<Self> {
+        NonZeroU64::new(p).map(|inner| Self {
             inner,
             _marker: PhantomData,
         })
     }
 
-    fn into_usize(&self) -> usize {
+    fn into_u64(&self) -> u64 {
         self.inner.get()
     }
 
-    fn tag(&self) -> u32 {
-        (self.inner.get() >> 32) as u32
+    fn tag(&self) -> NonZeroU32 {
+        let tag = (self.inner.get() >> 32) as u32;
+        debug_assert_ne!(0, tag, "broken non-zero invariant");
+        unsafe { NonZeroU32::new_unchecked(tag) }
     }
 
     fn incr_tag(&mut self) {
-        let tag = self.tag().wrapping_add(1);
+        let maybe_zero_tag = self.tag().get().wrapping_add(1);
+        let tag = NonZeroU32::new(maybe_zero_tag).unwrap_or(initial_tag_value());
         let offset = self.offset();
 
         *self = unsafe { Ptr::from_parts(tag, offset) };
@@ -147,16 +179,23 @@ impl<T> Ptr<T> {
         self.inner.get() as i32
     }
 
+    #[cfg(target_arch = "x86_64")]
     fn as_raw(&self) -> NonNull<T> {
         unsafe {
             NonNull::new_unchecked(
-                (anchor::<T>() as *mut u8).offset(self.offset() as isize) as *mut T
+                (anchor::<T>(None) as isize).wrapping_add(self.offset() as isize) as *mut T,
             )
         }
     }
 
+    #[cfg(target_arch = "x86")]
+    fn as_raw(&self) -> NonNull<T> {
+        unsafe { NonNull::new_unchecked(self.offset() as *mut T) }
+    }
+
     pub fn dangling() -> Self {
-        unsafe { Self::from_parts(0, 1) }
+        // `anchor()` returns a well-aligned pointer so an offset of 0 will also produce a well-aligned pointer
+        unsafe { Self::from_parts(initial_tag_value(), 0) }
     }
 
     pub unsafe fn as_ref(&self) -> &T {
@@ -165,14 +204,14 @@ impl<T> Ptr<T> {
 }
 
 struct Atomic<T> {
-    inner: AtomicUsize,
+    inner: AtomicU64,
     _marker: PhantomData<*mut T>,
 }
 
 impl<T> Atomic<T> {
     const fn null() -> Self {
         Self {
-            inner: AtomicUsize::new(0),
+            inner: AtomicU64::new(0),
             _marker: PhantomData,
         }
     }
@@ -186,17 +225,17 @@ impl<T> Atomic<T> {
     ) -> Result<(), Option<Ptr<T>>> {
         self.inner
             .compare_exchange_weak(
-                current.map(|p| p.into_usize()).unwrap_or(0),
-                new.map(|p| p.into_usize()).unwrap_or(0),
+                current.map(|p| p.into_u64()).unwrap_or(0),
+                new.map(|p| p.into_u64()).unwrap_or(0),
                 succ,
                 fail,
             )
             .map(drop)
-            .map_err(Ptr::from_usize)
+            .map_err(Ptr::from_u64)
     }
 
     fn load(&self, ord: Ordering) -> Option<Ptr<T>> {
-        NonZeroUsize::new(self.inner.load(ord)).map(|inner| Ptr {
+        NonZeroU64::new(self.inner.load(ord)).map(|inner| Ptr {
             inner,
             _marker: PhantomData,
         })
@@ -204,6 +243,6 @@ impl<T> Atomic<T> {
 
     fn store(&self, val: Option<Ptr<T>>, ord: Ordering) {
         self.inner
-            .store(val.map(|p| p.into_usize()).unwrap_or(0), ord)
+            .store(val.map(|p| p.into_u64()).unwrap_or(0), ord)
     }
 }