+++ src/pool/mod.rs
@@ -2,7 +2,7 @@
 //!
 //! NOTE: This module is not available on targets that do *not* support CAS operations, e.g. ARMv6-M
 //!
-//! (\*) Currently, the implementation is only lock-free *and* `Sync` on ARMv7-{A,R,M} & ARMv8-M
+//! (\*) Currently, the implementation is only lock-free *and* `Sync` on ARMv6, ARMv7-{A,R,M} & ARMv8-M
 //! devices
 //!
 //! # Examples
@@ -153,7 +153,8 @@
 //!
 //! # x86_64 support / limitations
 //!
-//! *NOTE* `Pool` is only `Sync` on `x86_64` if the Cargo feature "x86-sync-pool" is enabled
+//! *NOTE* `Pool` is only `Sync` on `x86_64` and `x86` (`i686`) if the Cargo feature "x86-sync-pool"
+//! is enabled
 //!
 //! x86_64 support is a gamble. Yes, a gamble. Do you feel lucky enough to use `Pool` on x86_64?
 //!
@@ -203,6 +204,10 @@
 //!
 //! ## x86_64 Limitations
 //!
+//! *NOTE* this limitation does not apply to `x86` (32-bit address space). If you run into this
+//! issue, on an x86_64 processor try running your code compiled for `x86`, e.g. `cargo run --target
+//! i686-unknown-linux-musl`
+//!
 //! Because stack nodes must be located within +- 2 GB of the hidden `ANCHOR` variable, which
 //! lives in the `.bss` section, `Pool` may not be able to manage static references created using
 //! `Box::leak` -- these heap allocated chunks of memory may live in a very different address space.
@@ -233,15 +238,18 @@ use core::{
     marker::PhantomData,
     mem::MaybeUninit,
     ops::{Deref, DerefMut},
-    ptr,
+    ptr::{self, NonNull},
 };
 
 pub use stack::Node;
 use stack::{Ptr, Stack};
 
 pub mod singleton;
-#[cfg_attr(target_arch = "x86_64", path = "cas.rs")]
-#[cfg_attr(not(target_arch = "x86_64"), path = "llsc.rs")]
+#[cfg_attr(any(target_arch = "x86_64", target_arch = "x86"), path = "cas.rs")]
+#[cfg_attr(
+    not(any(target_arch = "x86_64", target_arch = "x86")),
+    path = "llsc.rs"
+)]
 mod stack;
 
 /// A lock-free memory pool
@@ -255,11 +263,15 @@ pub struct Pool<T> {
 
 // NOTE(any(test)) makes testing easier (no need to enable Cargo features for testing)
 #[cfg(any(
+    armv6m,
     armv7a,
     armv7r,
     armv7m,
     armv8m_main,
-    all(target_arch = "x86_64", feature = "x86-sync-pool"),
+    all(
+        any(target_arch = "x86_64", target_arch = "x86"),
+        feature = "x86-sync-pool"
+    ),
     test
 ))]
 unsafe impl<T> Sync for Pool<T> {}
@@ -283,6 +295,11 @@ impl<T> Pool<T> {
     /// *NOTE:* This method does *not* have bounded execution time because it contains a CAS loop
     pub fn alloc(&self) -> Option<Box<T, Uninit>> {
         if mem::size_of::<T>() == 0 {
+            // NOTE because we return a dangling pointer to a NODE, which has non-zero size
+            // even when T is a ZST, in this case we need to make sure we
+            // - don't do pointer arithmetic on this pointer
+            // - dereference that offset-ed pointer as a ZST
+            // because miri doesn't like that
             return Some(Box {
                 node: Ptr::dangling(),
                 _state: PhantomData,
@@ -309,8 +326,14 @@ impl<T> Pool<T> {
         S: 'static,
     {
         if TypeId::of::<S>() == TypeId::of::<Init>() {
+            let p = if mem::size_of::<T>() == 0 {
+                // any pointer will do to invoke the destructor of a ZST
+                NonNull::dangling().as_ptr()
+            } else {
+                unsafe { value.node.as_ref().data.get() }
+            };
             unsafe {
-                ptr::drop_in_place(value.node.as_ref().data.get());
+                ptr::drop_in_place(p);
             }
         }
 
@@ -328,13 +351,12 @@ impl<T> Pool<T> {
     ///
     /// This method returns the number of *new* blocks that can be allocated.
     pub fn grow(&self, memory: &'static mut [u8]) -> usize {
-        let sz = mem::size_of::<Node<T>>();
-
-        if sz == 0 {
-            // SZT use no memory so a pool of SZT always has maximum capacity
+        if mem::size_of::<T>() == 0 {
+            // ZST use no memory so a pool of ZST always has maximum capacity
             return usize::max_value();
         }
 
+        let sz = mem::size_of::<Node<T>>();
         let mut p = memory.as_mut_ptr();
         let mut len = memory.len();
 
@@ -355,19 +377,20 @@ impl<T> Pool<T> {
         let mut n = 0;
         while len >= sz {
             match () {
-                #[cfg(target_arch = "x86_64")]
+                #[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
                 () => {
                     if let Some(p) = Ptr::new(p as *mut _) {
                         self.stack.push(p);
+                        n += 1;
                     }
                 }
 
-                #[cfg(not(target_arch = "x86_64"))]
+                #[cfg(not(any(target_arch = "x86_64", target_arch = "x86")))]
                 () => {
                     self.stack.push(unsafe { Ptr::new_unchecked(p as *mut _) });
+                    n += 1;
                 }
             }
-            n += 1;
 
             p = unsafe { p.add(sz) };
             len -= sz;
@@ -392,14 +415,14 @@ impl<T> Pool<T> {
         let cap = nodes.len();
         for p in nodes {
             match () {
-                #[cfg(target_arch = "x86_64")]
+                #[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
                 () => {
                     if let Some(p) = Ptr::new(p) {
                         self.stack.push(p);
                     }
                 }
 
-                #[cfg(not(target_arch = "x86_64"))]
+                #[cfg(not(any(target_arch = "x86_64", target_arch = "x86")))]
                 () => self.stack.push(core::ptr::NonNull::from(p)),
             }
         }
@@ -416,8 +439,14 @@ pub struct Box<T, STATE = Init> {
 impl<T> Box<T, Uninit> {
     /// Initializes this memory block
     pub fn init(self, val: T) -> Box<T, Init> {
-        unsafe {
-            ptr::write(self.node.as_ref().data.get(), val);
+        if mem::size_of::<T>() == 0 {
+            // no memory operation needed for ZST
+            // BUT we want to avoid calling `val`s destructor
+            mem::forget(val)
+        } else {
+            unsafe {
+                ptr::write(self.node.as_ref().data.get(), val);
+            }
         }
 
         Box {
@@ -472,13 +501,23 @@ impl<T> Deref for Box<T> {
     type Target = T;
 
     fn deref(&self) -> &T {
-        unsafe { &*self.node.as_ref().data.get() }
+        if mem::size_of::<T>() == 0 {
+            // any pointer will do for ZST
+            unsafe { &*NonNull::dangling().as_ptr() }
+        } else {
+            unsafe { &*self.node.as_ref().data.get() }
+        }
     }
 }
 
 impl<T> DerefMut for Box<T> {
     fn deref_mut(&mut self) -> &mut T {
-        unsafe { &mut *self.node.as_ref().data.get() }
+        if mem::size_of::<T>() == 0 {
+            // any pointer will do for ZST
+            unsafe { &mut *NonNull::dangling().as_ptr() }
+        } else {
+            unsafe { &mut *self.node.as_ref().data.get() }
+        }
     }
 }