+++ src/sealed.rs
@@ -22,6 +22,36 @@ pub mod binary_heap {
     }
 }
 
+/// Sealed traits and implementations for `LinkedList`
+pub mod sorted_linked_list {
+    use crate::sorted_linked_list::{Max, Min};
+    use core::cmp::Ordering;
+
+    /// The linked list kind: min-list or max-list
+    pub unsafe trait Kind {
+        #[doc(hidden)]
+        fn ordering() -> Ordering;
+    }
+
+    unsafe impl Kind for Min {
+        fn ordering() -> Ordering {
+            Ordering::Less
+        }
+    }
+
+    unsafe impl Kind for Max {
+        fn ordering() -> Ordering {
+            Ordering::Greater
+        }
+    }
+}
+
+#[allow(dead_code)]
+#[allow(path_statements)]
+pub(crate) const fn smaller_than<const N: usize, const MAX: usize>() {
+    Assert::<N, MAX>::LESS;
+}
+
 #[allow(dead_code)]
 #[allow(path_statements)]
 pub(crate) const fn greater_than_0<const N: usize>() {
@@ -32,6 +62,13 @@ pub(crate) const fn greater_than_0<const N: usize>() {
 #[allow(path_statements)]
 pub(crate) const fn greater_than_1<const N: usize>() {
     Assert::<N, 1>::GREATER;
+}
+
+#[allow(dead_code)]
+#[allow(path_statements)]
+pub(crate) const fn power_of_two<const N: usize>() {
+    Assert::<N, 0>::GREATER;
+    Assert::<N, 0>::POWER_OF_TWO;
 }
 
 #[allow(dead_code)]
@@ -57,4 +94,7 @@ impl<const L: usize, const R: usize> Assert<L, R> {
 
     /// Const assert hack
     pub const LESS: usize = R - L - 1;
+
+    /// Const assert hack
+    pub const POWER_OF_TWO: usize = 0 - (L & (L - 1));
 }