+++ CHANGELOG.md
@@ -7,7 +7,53 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-## [v0.7.3] - 2021-07-1
+## [v0.7.7] - 2021-09-22
+
+### Fixed
+
+- Fixed so `Pool` is `Sync` on ARMv6
+
+## [v0.7.6] - 2021-09-21
+
+### Added
+
+- Added `ArcPool`
+- Added `Debug` impl for `Deque`
+
+### Fixed
+
+- ZSTs in `Pool` now works correctly
+- Some MIRI errors were resolved
+- Allow `pool!` on thumbv6
+- Fixed possible UB in `Pool` on x86
+
+## [v0.7.5] - 2021-08-16
+
+### Added
+
+- Added `SortedLinkedList`
+- Added `Vec::is_empty`, one does not need to go through a slice anymore
+
+### Changed
+
+- `Vec::pop_unchecked` is now public
+
+## [v0.7.4] - 2021-08-06
+
+### Added
+
+- Implement `Default` for `MpMcQueue`, `Queue` and `HistoryBuffer`
+- Implement `PartialOrd` and `Ord` for `Vec` and `String`
+
+### Fixed
+
+- Fixed comments in SPSC
+
+## [v0.7.3] - 2021-07-01
+
+### Added
+
+- Added `Deque`
 
 ### Changed
 
@@ -371,7 +417,11 @@ architecture.
 
 - Initial release
 
-[Unreleased]: https://github.com/japaric/heapless/compare/v0.7.3...HEAD
+[Unreleased]: https://github.com/japaric/heapless/compare/v0.7.7...HEAD
+[v0.7.7]: https://github.com/japaric/heapless/compare/v0.7.6...v0.7.7
+[v0.7.6]: https://github.com/japaric/heapless/compare/v0.7.5...v0.7.6
+[v0.7.5]: https://github.com/japaric/heapless/compare/v0.7.4...v0.7.5
+[v0.7.4]: https://github.com/japaric/heapless/compare/v0.7.3...v0.7.4
 [v0.7.3]: https://github.com/japaric/heapless/compare/v0.7.2...v0.7.3
 [v0.7.2]: https://github.com/japaric/heapless/compare/v0.7.1...v0.7.2
 [v0.7.1]: https://github.com/japaric/heapless/compare/v0.7.0...v0.7.1